backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
google-generativeai==0.3.0
pypdf>=3.0.0
pydantic==2.5.0
python-multipart==0.0.6
numpy==1.24.3
scikit-learn==1.3.2
sentence-transformers>=2.3.0
rapidfuzz==3.4.0
requests==2.31.0
Werkzeug==3.0.1
